feat: add MCP tools, pipeline validation, and multi-query RRF
Major Features:
- Complete MCP (Model Context Protocol) implementation with 8 tools
- Pipeline contract validation framework for API consistency
- Multi-Query RRF pipeline with reciprocal rank fusion
- Enhanced test fixtures with automatic schema migration

MCP Tools (Feature 043):
- 6 RAG pipeline tools (basic, basic_rerank, crag, graphrag, multi_query_rrf, hybrid_graphrag)
- 2 utility tools (health_check, list_tools)
- MCP schema moved to package for public distribution

Pipeline Validation (Feature 047):
- Standardized API contracts across all pipelines
- 100% LangChain & RAGAS compatibility validation
- Automated contract testing with validators
- 34 new contract and integration tests

Multi-Query RRF (Feature 048):
- Query expansion with multiple perspectives
- Reciprocal rank fusion for result merging
- Integration with RAGAS evaluation framework

Testing Improvements:
- Working integration test fixtures with schema migration
- Programmatic data loading for repeatable tests
- 11 MCP integration tests passing
- Fixed RAGAS evaluation for all 6 pipelines

Bug Fixes:
- MCP schema path moved from specs/ to iris_rag/mcp/
- RAGAS evaluation includes multi_query_rrf
- Test fixtures handle VECTOR schema migration
- Test expectations corrected for 5-document fixture

+++ iris_rag/pipelines/crag.py
@@ -494,7 +494,7 @@ class CRAGPipeline(RAGPipeline):
                 SELECT TOP {top_k}
                     doc_id,
                     chunk_text,
-                    VECTOR_COSINE(chunk_embedding, TO_VECTOR(?, DOUBLE, 384)) as similarity_score
+                    VECTOR_COSINE(chunk_embedding, TO_VECTOR(?, FLOAT, 384)) as similarity_score
                 FROM RAG.DocumentChunks
                 WHERE chunk_embedding IS NOT NULL
                 ORDER BY similarity_score DESC
@@ -606,7 +606,7 @@ class CRAGPipeline(RAGPipeline):
                     SELECT TOP {top_k * 2}
                         doc_id,
                         text_content,
-                        VECTOR_COSINE(embedding, TO_VECTOR(?, DOUBLE, 384)) as similarity_score
+                        VECTOR_COSINE(embedding, TO_VECTOR(?, FLOAT, 384)) as similarity_score
                     FROM RAG.SourceDocuments
                     WHERE embedding IS NOT NULL
                     ORDER BY similarity_score DESC
